+++ weblate/trans/models/label.py
@@ -23,6 +23,8 @@ from django.utils.html import escape
 from django.utils.safestring import mark_safe
 from django.utils.translation import gettext_lazy
 
+from weblate.utils.colors import COLOR_CHOICES
+
 
 class Label(models.Model):
     project = models.ForeignKey("Project", on_delete=models.deletion.CASCADE)
@@ -30,40 +32,7 @@ class Label(models.Model):
     color = models.CharField(
         verbose_name=gettext_lazy("Color"),
         max_length=30,
-        choices=(
-            # Translators: Name of a color
-            ("navy", gettext_lazy("Navy")),
-            # Translators: Name of a color
-            ("blue", gettext_lazy("Blue")),
-            # Translators: Name of a color
-            ("aqua", gettext_lazy("Aqua")),
-            # Translators: Name of a color
-            ("teal", gettext_lazy("Teal")),
-            # Translators: Name of a color
-            ("olive", gettext_lazy("Olive")),
-            # Translators: Name of a color
-            ("green", gettext_lazy("Green")),
-            # Translators: Name of a color
-            ("lime", gettext_lazy("Lime")),
-            # Translators: Name of a color
-            ("yellow", gettext_lazy("Yellow")),
-            # Translators: Name of a color
-            ("orange", gettext_lazy("Orange")),
-            # Translators: Name of a color
-            ("red", gettext_lazy("Red")),
-            # Translators: Name of a color
-            ("maroon", gettext_lazy("Maroon")),
-            # Translators: Name of a color
-            ("fuchsia", gettext_lazy("Fuchsia")),
-            # Translators: Name of a color
-            ("purple", gettext_lazy("Purple")),
-            # Translators: Name of a color
-            ("black", gettext_lazy("Black")),
-            # Translators: Name of a color
-            ("gray", gettext_lazy("Gray")),
-            # Translators: Name of a color
-            ("silver", gettext_lazy("Silver")),
-        ),
+        choices=COLOR_CHOICES,
         blank=False,
         default=None,
     )