+++ weblate/wladmin/sites.py
@@ -41,6 +41,8 @@ from weblate.checks.admin import CheckAdmin
 from weblate.checks.models import Check
 from weblate.fonts.admin import FontAdmin, FontGroupAdmin
 from weblate.fonts.models import Font, FontGroup
+from weblate.glossary.admin import GlossaryAdmin, TermAdmin
+from weblate.glossary.models import Glossary, Term
 from weblate.lang.admin import LanguageAdmin
 from weblate.lang.models import Language
 from weblate.memory.admin import MemoryAdmin
@@ -54,7 +56,6 @@ from weblate.trans.admin import (
     ComponentAdmin,
     ComponentListAdmin,
     ContributorAgreementAdmin,
-    DictionaryAdmin,
     ProjectAdmin,
     SuggestionAdmin,
     TranslationAdmin,
@@ -67,7 +68,6 @@ from weblate.trans.models import (
     Component,
     ComponentList,
     ContributorAgreement,
-    Dictionary,
     Project,
     Suggestion,
     Translation,
@@ -118,6 +118,7 @@ class WeblateAdminSite(AdminSite):
         self.register(Announcement, AnnouncementAdmin)
         self.register(ComponentList, ComponentListAdmin)
         self.register(ContributorAgreement, ContributorAgreementAdmin)
+        self.register(Glossary, GlossaryAdmin)
 
         # Show some controls only in debug mode
         if settings.DEBUG:
@@ -126,7 +127,7 @@ class WeblateAdminSite(AdminSite):
             self.register(Suggestion, SuggestionAdmin)
             self.register(Comment, CommentAdmin)
             self.register(Check, CheckAdmin)
-            self.register(Dictionary, DictionaryAdmin)
+            self.register(Term, TermAdmin)
             self.register(Change, ChangeAdmin)
 
         # Billing